cache: add basic quote query caching with sqlite
I couldn't have done it so quickly without gorp, wonderful little library. A
change I had to make because of gorp was that I had to flatten the
fquery.Result struct, as it doesn't deal well with nested structs (yet?).
But that's a small price to pay for seamless persistence, fantastic!