yql: support queries with 1 symbol
It appears the resultant is slightly different from the case with multiple
symbols. I.e.: the in the case of multiple symbols, the quotes get returned
in an array. Otherwise, the quote is a free-standing JSON object, not an
array. This change introduces an alternative Go struct to deal with the
issue.